==> Scripts/A_Preprocessing/05_compile_sample_metadat.R <==
# ==============================================================================
# A/05_compile_metadat.R
# parsing, cleaning, merge metadata into unified fmt (redcap, methyl qc, etc)
# trial outcomes coded as e.g., outcome_bmi_0 (baseline) or outcome_bmi_3 (3-mo)
#                      and outcomedelta_bmi_30 (3-mo change), _60 (6-mo change)
# ==============================================================================




# code not shown
# omitted for irb/limit participant deidentification
